Step4: Move the pilot canvas and app shell into place
The 2D drawing surface (R2DWidget, RWorldRenderer2d, DrawTool, and
RTextOverlay, which the canvas renders its text with) moves into
canvas/, and the app shell (the RManager window hub plus the menu,
action, keymap, and shortcut-manager machinery) moves into app/.

The shell moves last on purpose: RManager is the hub that includes the
most siblings, so its include rewrite lands against a tree where every
other neighborhood already has its final path.  The Qt-free keymap
core stays beside the menu and shortcuts it drives, and the gtest list
updates its direct reference to keymap.cpp.  With this step the flat
directory is gone; only the umbrella pilot.hpp and the wrap_pilot
binding unit remain at the package root.  No logic changes.
