+++ cpp/solvcon/pilot/app/RMenuModel.cpp
@@ -3,7 +3,7 @@
  * BSD 3-Clause License, see COPYING
  */
 
-#include <solvcon/pilot/RMenuModel.hpp> // Must be the first include.
+#include <solvcon/pilot/app/RMenuModel.hpp> // Must be the first include.
 
 #include <limits>
