+++ cpp/solvcon/pilot/canvas/DrawTool.hpp
