+++ cpp/solvcon/pilot/canvas/R2DWidget.cpp
@@ -3,7 +3,7 @@
  * BSD 3-Clause License, see COPYING
  */
 
-#include <solvcon/pilot/R2DWidget.hpp> // Must be the first include.
+#include <solvcon/pilot/canvas/R2DWidget.hpp> // Must be the first include.
 
 #include <array>
 #include <cmath>
