+++ cpp/solvcon/pilot/canvas/RTextOverlay.hpp
