+++ cpp/solvcon/pilot/canvas/RWorldRenderer2d.hpp
