+++ cpp/solvcon/pilot/pilot.hpp
@@ -15,7 +15,7 @@
 #include <solvcon/pilot/common/common_detail.hpp> // Must be the first include.
 
 #include <solvcon/pilot/visual/RDomainWidget.hpp>
-#include <solvcon/pilot/RManager.hpp>
+#include <solvcon/pilot/app/RManager.hpp>
 #include <solvcon/pilot/console/RPythonConsoleDockWidget.hpp>
 
 // vim: set ff=unix fenc=utf8 et sw=4 ts=4 sts=4:
